refactor(filter-category): make it a molecule instead
